⚙️ Experimenting with a custom module
Pretty clear advantages to this, but it is also a lot of scaffolding for
a pretty simple outcome. There will be a line between when to use this.
I should learn a little more about overlays also, as it may impact how I
use modules.

An obvious module, would be hyprland. There is a LOT in there and I
already have a few things I need to allow configuration between
machines. EG waybar modules on Desktop vs Laptop (battery, GPU)

🤷 we shall see!