feat: Add script for creating new git repositories
Add a script to facilitate the creation of new git repositories.
This is useful for moving aspects of the current repository to
separate repositories. The script includes functionality for
creating the directory structure, initializing a git repository,
and setting up a new GitHub repository.

A future version should likely allow for creating a private repository.