feat: Move dunst to a systemd service
Implement dunst as a systemd service to improve its lifecycle
management. I am not sure if this is really needed. It seems dbus
automatically starts it if needed. Something about requests for
'org.freedesktop.Notifications', dbus dunst will be automatically
starte dunst if it is not already running.

But anyway, I believe this change enhances the reliability and
efficiency of the notification service