feat(vim): Call cd in Oil to goto directory
First look at cross pane commands in Tmux. I always work in a 1 up 1
down tmux layout. Especially when using vim. See my tmuxifier. So this
is a really simple way to use that knowledge to work accross panes.

It isnt smart, it isn't very extensible. But it is a start.