feat: Experimenting with zen-browser settings
My goal here is to have it setup enought that I dont need a Firefox
account. So bookmarks, extentions, theming and more. Ideally I can
easily have a work and personal profile that are very similar visually,
but different bookmarks and logins. Maybe different themes to
differenciate.

As with the who repo, it is because I can, rather should 🤷