feat: add initial webapp functionality
The ability to launch any webpage as an "app", ie, its own window with
without any browser UI beyond the page content.

This was infuriating to do. I use Firefox as my main browser, always
have (tried zen for a little while), so I didn't want to install another
browser to just do this. Here were the problems:

Firefox does not have an "app" mode. With chromium you simply pass
`--app=<url>` and you are away. I may have been able to do this with a
custom Profile, but it seemed like overkill, but given the following I
may go back and try it. There is `--kiosk` which is similar but
fullscreens the window also, Hyprland can in theory get around this by
having windowrules, but..

Firefox sets it's window title some time after it has launched,
initially it is simple "firefox" or similar. Windowrules exec once at
launch, meaning we would need a mechanism to recheck the title after
some time and then move it. This is dumb, and I didn't go there. In
general this means there is no way to set Title specific window rules on
Firefox windows. Meaning you are limited to setting rules for Firefox as
a whole.

So what if we just be very explicit and tell Firefox as it launches
where to go and what to do...

The CLI command `firefox` is not the process that runs the window you
see and interact with. So with `--kiosk` and having Hyprland lie about
its fullscreenness¹, but this requires hyprctl to have the PID of the
window, NOT the binary that launches the window. I could not see any way
to get it reliably either. Firefox and other browsers will spawn any
number of processes, apparently multiple windows can be tied to a single
PID!²

So I decided to just bite the bullet and install chromium. But I like to
not kneel before our techno-feudal overlords too much, so went with
`ungoogled-chromium` instead.

Problems with this approach:

* An additional pkgs to install and manage. Not the end of the world.
  Nix has allowed me to be very specific about what I have installed, so
  I feel confident there aren't many things I don't need lying around.
* No shared sessions between main browser and webapps. This is the big
  one really, I will need to sign in to everything again even if I am
  already signed in on Firefox.

---

[1] - Hyprland has a neat trick where you can essentially tell the
      window it's in fullscreen when it is actually windowed.
      `fullscreenstate <internal mode> <client mode>`, so set internal
      (hyprland) to windowed `0`, set client to `2` fullscreen. This would
      be called like so `hyprcyl dispatch exec "[fullscreenstate 0 2]
      firefox"`
[2] - I feel I may be able to get around that last part with come
      namespaces or what not, would `uwsm app --` solve that??