fix: solve cowardly precommit by using it in glbls
```
[ERROR] Cowardly refusing to install hooks with `core.hooksPath` set.
hint: `git config --unset-all core.hooksPath`
```
I was getting this error. This is because pre-commit looks for an
existing hook and if it exists it doesn't install. This only shows if it
has not already installed itself in to .git/hooks/pre-commit.

I removed this repo and recloned it in a worktree setup, which caused
pre-commit to need to reinstall.

https://warman.io/blog/2024/01/global-pre-commit/ was a good reference
for this. The main change was to simply capture the exit code on each
invocation of pre-commit and exit when they all have run. So I still run
local linters etc if I have !nocommit

Next step might be to look more at how pre-commit works and get it to
pass files to my no-commit script, will hopefully event be able to check
if the nocommit patch is staged, not just the whole file 🤷