refacotr: big Nix refactor
Pulling apart the master configuration.nix files. Probably controversial
but I never like a "master" file with all sorts of things in, I find it
overwhelming and hard to grok. Now that doesn't mean there are not
downsides to this. When I add my server I likely will need to revisit
"common", I touched on this slightly with the ssh.enable settings. I
expect I will need to re-enable this for a server.